+++ qa/restart-upgrade/src/test/java/org/opensearch/neuralsearch/bwc/KnnRadialSearchIT.java
@@ -61,6 +61,7 @@ public class KnnRadialSearchIT extends AbstractRestartUpgradeRestTestCase {
             0.01f,
             null,
             null,
+            null,
             null
         );
         Map<String, Object> responseWithMinScoreQuery = search(getIndexNameForTest(), neuralQueryBuilderWithMinScoreQuery, 1);
@@ -73,6 +74,7 @@ public class KnnRadialSearchIT extends AbstractRestartUpgradeRestTestCase {
             modelId,
             null,
             100000f,
+            null,
             null,
             null,
             null,