+++ src/test/java/org/opensearch/neuralsearch/processor/NormalizationProcessorIT.java
@@ -97,6 +97,7 @@ public class NormalizationProcessorIT extends BaseNeuralSearchIT {
                 null,
                 null,
                 null,
+                null,
                 null
             );
             TermQueryBuilder termQueryBuilder = QueryBuilders.termQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT3);
@@ -148,6 +149,7 @@ public class NormalizationProcessorIT extends BaseNeuralSearchIT {
                 null,
                 null,
                 null,
+                null,
                 null
             );
             TermQueryBuilder termQueryBuilder = QueryBuilders.termQuery(TEST_TEXT_FIELD_NAME_1, TEST_QUERY_TEXT3);
@@ -184,6 +186,7 @@ public class NormalizationProcessorIT extends BaseNeuralSearchIT {
                 "",
                 modelId,
                 6,
+                null,
                 null,
                 null,
                 null,