+++ src/testFixtures/java/org/opensearch/neuralsearch/BaseNeuralSearchIT.java
@@ -203,6 +203,7 @@ public abstract class BaseNeuralSearchIT extends OpenSearchSecureRestTestCase {
             isComplete = checkComplete(taskQueryResult);
             Thread.sleep(DEFAULT_TASK_RESULT_QUERY_INTERVAL_IN_MILLISECOND);
         }
+        assertTrue(isComplete);
     }
 
     /**
@@ -528,7 +529,7 @@ public abstract class BaseNeuralSearchIT extends OpenSearchSecureRestTestCase {
         Map<String, String> requestParams,
         List<Object> aggs
     ) {
-        return search(index, queryBuilder, rescorer, resultSize, requestParams, aggs, null, null, false, null);
+        return search(index, queryBuilder, rescorer, resultSize, requestParams, aggs, null, null, false, null, 0);
     }
 
     @SneakyThrows
@@ -542,10 +543,11 @@ public abstract class BaseNeuralSearchIT extends OpenSearchSecureRestTestCase {
         QueryBuilder postFilterBuilder,
         List<SortBuilder<?>> sortBuilders,
         boolean trackScores,
-        List<Object> searchAfter
+        List<Object> searchAfter,
+        int from
     ) {
         XContentBuilder builder = XContentFactory.jsonBuilder().startObject();
-
+        builder.field("from", from);
         if (queryBuilder != null) {
             builder.field("query");
             queryBuilder.toXContent(builder, ToXContent.EMPTY_PARAMS);
@@ -1199,6 +1201,8 @@ public abstract class BaseNeuralSearchIT extends OpenSearchSecureRestTestCase {
         Request waitForGreen = new Request("GET", "/_cluster/health");
         waitForGreen.addParameter("wait_for_nodes", numOfNodes);
         waitForGreen.addParameter("wait_for_status", "green");
+        waitForGreen.addParameter("cluster_manager_timeout", "60s");
+        waitForGreen.addParameter("timeout", "60s");
         client().performRequest(waitForGreen);
     }